requirements.txt
pyserial>=2.0
Flask>=1.0.2
Flask-RESTful>=0.3.6
simplejson>=3.16.0
